requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopage==0.5.0
backcall==0.2.0
black==19.10b0
blis==0.7.7
cachetools==5.1.0
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.4
colorlog==6.6.0
cymem==2.0.6
decorator==5.1.1
docker-pycreds==0.4.0
executing==0.8.3
FastNLP @ git+https://gitee.com/LouChao98/fastNLP.git
filelock==3.7.0
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.46.3
huggingface-hub==0.6.0
hydra-core==1.1.0
hydra-joblib-launcher==1.2.0
hydra-optuna-sweeper==1.2.0
idna==3.3
importlib-metadata==4.11.4
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
llvmlite==0.38.1
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mkl-fft==1.3.0
mkl-random==1.0.2
mkl-service==2.3.0
multidict==6.0.2
murmurhash==1.0.7
mypy-extensions==0.4.3
nltk==3.7
numba==0.55.1
numpy==1.19.2
oauthlib==3.2.0
omegaconf==2.1.2
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
parso==0.8.3
pathspec==0.7.0
pathtools==0.1.2
pathy==0.6.1
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.4
preshed==3.0.6
prettytable==3.3.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytorch-lightning==1.4.9
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==59.5.0
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sparse==0.13.0
SQLAlchemy==1.4.36
srsly==2.4.3
stack-data==0.2.0
stevedore==3.5.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.0.16
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
torchmetrics==0.6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.19.2
typed-ast==1.4.3
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.16
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.37.1
yarl==1.7.2
zipp==3.8.0
